+++ lib/models/practice_type.dart
@@ -1,6 +1,13 @@
-class PracticeType {
-  const PracticeType({required this.title, required this.description});
+enum PracticeTypeId { binary, loopScout }
 
+class PracticeType {
+  const PracticeType({
+    required this.id,
+    required this.title,
+    required this.description,
+  });
+
+  final PracticeTypeId id;
   final String title;
   final String description;
 }